# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API OmniGest",
        "description": "Documentação da API Omnigest",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api"
        }
    ],
    "paths": {
...
        "/empresa": {
            "post": {
                "tags": [
                    "Empresa"
                ],
                "summary": "Cria uma nova empresa",
                "operationId": "465cc9f908a6fab2708eb338a284383c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "des_empresa_emp",
                                    "razao_social_empresa_emp",
                                    "cnpj_empresa_emp",
                                    "des_endereco_emp",
                                    "des_cidade_emp",
                                    "des_cep_emp",
                                    "des_tel_emp"
                                ],
                                "properties": {
                                    "des_empresa_emp": {
                                        "type": "string"
                                    },
                                    "razao_social_empresa_emp": {
                                        "type": "string"
                                    },
                                    "cnpj_empresa_emp": {
                                        "type": "string"
                                    },
                                    "des_endereco_emp": {
                                        "type": "string"
                                    },
                                    "des_cidade_emp": {
                                        "type": "string"
                                    },
                                    "des_cep_emp": {
                                        "type": "string"
                                    },
                                    "des_tel_emp": {
                                        "type": "string"
                                    },
                                    "lnk_whatsapp_emp": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "lnk_instagram_emp": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "lnk_facebook_emp": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "img_empresa_emp": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Empresa criada com sucesso"
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                }
            }
        },
        "/empresas": {
            "get": {
                "tags": [
                    "Empresa"
                ],
                "summary": "Lista todas as empresas",
                "operationId": "7d73e5b455f07ff5c2e3a39b08ddedbd",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Quantidade de registros por página",
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    },
                    {
                        "name": "filter",
                        "in": "query",
                        "description": "Filtro para busca",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page_number",
                        "in": "query",
                        "description": "Número da página",
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de empresas retornada com sucesso"
                    }
                }
            }
        },
        "/empresa/{id_empresa_emp}": {
            "put": {
                "tags": [
                    "Empresa"
                ],
                "summary": "Atualiza uma empresa existente",
                "operationId": "5ee05685815a03e812bebee8ffd2d4c1",
                "parameters": [
                    {
                        "name": "id_empresa_emp",
                        "in": "path",
                        "description": "ID da empresa",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "des_empresa_emp",
                                    "razao_social_empresa_emp",
                                    "cnpj_empresa_emp",
                                    "des_endereco_emp",
                                    "des_cidade_emp",
                                    "des_cep_emp",
                                    "des_tel_emp"
                                ],
                                "properties": {
                                    "des_empresa_emp": {
                                        "type": "string"
                                    },
                                    "razao_social_empresa_emp": {
                                        "type": "string"
                                    },
                                    "cnpj_empresa_emp": {
                                        "type": "string"
                                    },
                                    "des_endereco_emp": {
                                        "type": "string"
                                    },
                                    "des_cidade_emp": {
                                        "type": "string"
                                    },
                                    "des_cep_emp": {
                                        "type": "string"
                                    },
                                    "des_tel_emp": {
                                        "type": "string"
                                    },
                                    "lnk_whatsapp_emp": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "lnk_instagram_emp": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "lnk_facebook_emp": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "img_empresa_emp": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Empresa atualizada com sucesso"
                    },
                    "404": {
                        "description": "Empresa não encontrada"
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}